TT.baseplot(soiltexture)	R Documentation

Internal. Create an empty plot scene for a texture triangle.

Description

Create an empty plot where a texture triangle can be drawn with other secondary functions (frame, axis, ...). Also return the 'geo' parameters needed by these secondary functions.

Usage


TT.baseplot(geo = NULL, class.sys = "none", blr.clock = NULL, tlr.an = NULL, blr.tx = NULL, text.sum = NULL, base.css.ps.lim = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, text.tol = NULL, unit.ps = NULL, unit.tx = NULL, b.lim = NULL, l.lim = NULL, main = NULL, new.mar = NULL, bg = NULL, fg = NULL, col = NULL, cex.main = NULL, lang = NULL)

Arguments

geo

class.sys

blr.clock

tlr.an

blr.tx

text.sum

base.css.ps.lim

tri.sum.tst

tri.pos.tst

text.tol

unit.ps

unit.tx

b.lim

l.lim

main

new.mar

bg

fg

col

cex.main

lang

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]