TT.blr.ps.lim(soiltexture)	R Documentation

Internal. Create a tabular version of clay silt sand particle size limits.

Description

Create a tabular version of clay silt sand particle size limits.

Usage


TT.blr.ps.lim(blr.tx, css.ps.lim)

Arguments

blr.tx

css.ps.lim

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]