TT.classes(soiltexture)	R Documentation

Plot the texture classes polygons in a texture triangle plot.

Description

Plot the texture classes ploygons in an existing texture triangle plot. Draw the polygons and the labels inside each polygons.

Usage


TT.classes(geo, class.sys, tri.css.ps.lim = NULL, css.transf = NULL, text.transf.fun = NULL, trsf.add.opt1 = NULL, trsf.add.opt2 = NULL, text.tol = NULL, text.sum = NULL, base.css.ps.lim = NULL, blr.tx = NULL, blr.clock = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, bg = NULL, class.lab.col = NULL, class.p.bg.col = NULL, class.p.bg.hue = NULL, class.line.col = NULL, class.lty = NULL, class.lab.show = NULL, cex.lab = NULL, font.lab = NULL, family.op = NULL, lwd.axis = NULL, col.axis = NULL, new.centroid = TRUE)

Arguments

geo

class.sys

tri.css.ps.lim

css.transf

text.transf.fun

trsf.add.opt1

trsf.add.opt2

text.tol

text.sum

base.css.ps.lim

blr.tx

blr.clock

tri.sum.tst

tri.pos.tst

bg

class.lab.col

class.p.bg.col

class.p.bg.hue

class.line.col

class.lty

class.lab.show

cex.lab

font.lab

family.op

lwd.axis

col.axis

new.centroid 	 Single logical. If TRUE (default) the new method (Paul Bourke) is used to calculate the centroid. If FALSE the centroid is taken as the mean x and y coordinates of the vertices.

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]