TT.classes.tbl(soiltexture)	R Documentation

Returns the table of classes of a texture classification system.

Description

Returns the table of classes of a texture classification system. Returns the classes abbreviations, names and the vertices numbers that defines each class. Use TT.vertices.tbl() to retrieve the clay silt sand coordinates of the triangle classes vertices. See also TT.vertices.plot().

Usage


TT.classes.tbl(class.sys = "FAO50.TT", collapse = ", ")

Arguments

class.sys

collapse

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]