TT.data.test.X(soiltexture)	R Documentation

Test the validity of some soil texture data table (X particle size classes).

Description

Test the validity of some soil texture data table. (1) Test that it is a data.frame or matrix, (3) Test that there are no missing values, (4) that all values are >= 0, (5) That the sum of the X particle size class is >= 'text.sum'*(1-'text.tol') or <= 'text.sum'*(1+'text.tol'). Contrary to TT.data.test() no test are performed for the particle size classes and columns names, so 'tri.data' should only contains texture data, and nothing else.

Usage


TT.data.test.X(tri.data, text.sum = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL)

Arguments

tri.data

text.sum

text.tol

tri.sum.tst

tri.pos.tst

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]