TT.dataset(soiltexture)	R Documentation

Genetates a virtual cross correlated clay silt sand + Z dataset.

Description

Genetates a virtual cross correlated clay silt sand + Z dataset, where Z is a virtual 4th variable correlated to the texture.

Usage


TT.dataset(n, seed.val = NULL, css.names = NULL, text.sum = NULL)

Arguments

n

seed.val

css.names

text.sum

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]