TT.deg2rad(soiltexture)	R Documentation

Function to convert angle in degree to angle in radian.

Description

Function to convert angle in degree to angle in radian.

Usage


TT.deg2rad(A)

Arguments

A 	 Angle in Degrees

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]