TT.geo.get(soiltexture)	R Documentation

Internal. Retrieve and return the geometrical parameters from a list of parameter values (NULL or not).

Description

Retrieve and return the geometrical parameters from a list of parameter values (NULL or not).

Usage


TT.geo.get(class.sys = NULL, blr.clock = NULL, tlr.an = NULL, blr.tx = NULL, text.sum = NULL, base.css.ps.lim = NULL)

Arguments

class.sys

blr.clock

tlr.an

blr.tx

text.sum

base.css.ps.lim

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]