TT.grid(soiltexture)	R Documentation

Plot a grid at regular texture intervals inside an existing soil texture triangle.

Description

Plot a grid at regular texture intervals inside an existing soil texture triangle.

Usage


TT.grid(geo = geo, at = NULL, text.tol = NULL, text.sum = NULL, blr.clock = NULL, grid.col = NULL, grid.lty = NULL, lwd.axis = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, class.p.bg.col = NULL, class.p.bg.hue = NULL, frame.bg.col = NULL, bg = NULL, col.axis = NULL)

Arguments

geo

at

text.tol

text.sum

blr.clock

grid.col

grid.lty

lwd.axis

tri.sum.tst

tri.pos.tst

class.p.bg.col

class.p.bg.hue

frame.bg.col

bg

col.axis

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]