TT.lines(soiltexture)	R Documentation

Internal. Used to plot line elements of a texture plot axis ticks arrows etc.

Description

Used to plot line elements of a texture plot axis, ticks, arrows, etc.

Usage


TT.lines(geo = geo, at.1.s = TT.get("at"), at.2.s = 1 - TT.get("at"), at.3.s = 0, at.1.e = TT.get("at"), at.2.e = 0, at.3.e = 1 - TT.get("at"), text.tol = NULL, text.sum = NULL, blr.clock = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL)

Arguments

geo

at.1.s

at.2.s

at.3.s

at.1.e

at.2.e

at.3.e

text.tol

text.sum

blr.clock

tri.sum.tst

tri.pos.tst

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]