TT.mahalanobis(soiltexture)	R Documentation

Calculates the Mahalanobis distance between clay silt and sand.

Description

Function that calculated the Mahalanobis distance between clay silt and sand, on a regular x-y grid (back-transformed to Clay silt and sand for Mahalanobis calculation). The underlying function is mahalanobis() by R Development Core Team (2009)

Usage


TT.mahalanobis(geo, tri.data, css.names = NULL, text.tol = NULL, text.sum = NULL, blr.clock = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, set.par = FALSE, n = 25, center = NULL, cov.mat = NULL, inverted = FALSE, ..., alr = FALSE, divisorvar = 2)

Arguments

geo

tri.data

css.names

text.tol

text.sum

blr.clock

tri.sum.tst

tri.pos.tst

set.par

n

center

cov.mat

inverted

...

alr

divisorvar

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]