TT.phi2dia(soiltexture)	R Documentation

Convert a soil particle phi value into diameter dia [micro-meters].

Description

Convert a soil particle phi value into diameter dia [micro-meters]. See also TT.dia2phi(). dia = (2^-phi)*1000. Not used by the package.

Usage


TT.phi2dia(phi)

Arguments

phi

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]