TT.points(soiltexture)	R Documentation

Plot a soil texture data table as points on an existing texture plot.

Description

Plot a soil texture data table as points on an existing texture plot.

Usage


TT.points(tri.data, geo, css.names = NULL, z.name = NULL, base.css.ps.lim = NULL, dat.css.ps.lim = NULL, css.transf = NULL, text.transf.fun = NULL, trsf.add.opt1 = NULL, trsf.add.opt2 = NULL, text.tol = NULL, pch = NULL, fg = NULL, col = NULL, bg = NULL, cex = NULL, lwd = NULL, points.type = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, z.type = NULL, z.col.hue = NULL, z.cex.range = NULL, z.pch = NULL, text.sum = NULL, blr.clock = NULL, blr.tx = NULL)

Arguments

tri.data

geo

css.names

z.name

base.css.ps.lim

dat.css.ps.lim

css.transf

text.transf.fun

trsf.add.opt1

trsf.add.opt2

text.tol

pch

fg

col

bg

cex

lwd

points.type

tri.sum.tst

tri.pos.tst

z.type

z.col.hue

z.cex.range

z.pch

text.sum

blr.clock

blr.tx

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]