TT.text(soiltexture)	R Documentation

Plot text labels for each values of a soil texture data table on an existing texture plot.

Description

Plot text labels for each values of a soil texture data table on an existing texture plot.

Usage


TT.text(tri.data, geo, labels = NULL, css.names = NULL, base.css.ps.lim = NULL, dat.css.ps.lim = NULL, css.transf = NULL, text.transf.fun = NULL, trsf.add.opt1 = NULL, trsf.add.opt2 = NULL, text.tol = NULL, text.sum = NULL, blr.clock = NULL, fg = NULL, col = NULL, cex = NULL, font = NULL, family.op = NULL, adj = NULL, pos = NULL, offset = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, blr.tx = NULL)

Arguments

tri.data

geo

labels

css.names

base.css.ps.lim

dat.css.ps.lim

css.transf

text.transf.fun

trsf.add.opt1

trsf.add.opt2

text.tol

text.sum

blr.clock

fg

col

cex

font

family.op

adj

pos

offset

tri.sum.tst

tri.pos.tst

blr.tx

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]