TT.text.transf(soiltexture)	R Documentation

Log-linear transformation of a soil texture data table between 2 particle size systems (3 classes).

Description

Log-linear transformation of a soil texture data table ('tri.data') from one particle size system ('dat.css.ps.lim') into another ('base.css.ps.lim'). Only 3 particle size classes allowed. See TT.text.transf.X for transformation involving more than 3 particle classes. 'tri.data' may contain other variables (not in 'css.names'). They are returned unchanged with the transformed texture data.

Usage


TT.text.transf(tri.data, base.css.ps.lim, dat.css.ps.lim, css.names = NULL, blr.tx = NULL, text.sum = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL, trsf.add.opt1 = NULL, trsf.add.opt2 = NULL)

Arguments

tri.data

base.css.ps.lim

dat.css.ps.lim

css.names

blr.tx

text.sum

text.tol

tri.sum.tst

tri.pos.tst

trsf.add.opt1

trsf.add.opt2

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]