TT.text.transf.X(soiltexture)	R Documentation

Log-linear transformation of a soil texture data table between 2 particle size systems (X classes).

Description

Log-linear transformation of a soil texture data table ('tri.data') from one particle size system ('dat.css.ps.lim') into another ('base.css.ps.lim'). No limit in the number of partile size classes in the inputed and outputed texture tables. See TT.text.transf for transformation involving only 3 particle classes. 'tri.data' can only contain texture data.

Usage


TT.text.transf.X(tri.data, base.ps.lim, dat.ps.lim, text.sum = NULL, text.tol = NULL, tri.sum.tst = NULL, tri.pos.tst = NULL)

Arguments

tri.data

base.ps.lim

dat.ps.lim

text.sum

text.tol

tri.sum.tst

tri.pos.tst

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]