TT.vertices.plot(soiltexture)	R Documentation

Plot the vertices of a texture classification system.

Description

Plot the vertices of a texture classification system, on top of an already drawn texture triangle plot. Also plot the vertices numbers. See TT.vertices.tbl() and TT.classes.tbl() for a non graphic, tabular equivalent of the plot.

Usage


TT.vertices.plot(geo, class.sys = "FAO50.TT", fg = NULL, col = NULL, cex = NULL, font = NULL, family.op = NULL, adj = NULL, pos = NULL, offset = NULL, blr.tx = NULL, text.sum = NULL, blr.clock = NULL)

Arguments

geo

class.sys

fg

col

cex

font

family.op

adj

pos

offset

blr.tx

text.sum

blr.clock

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]