TT.vertices.tbl(soiltexture)	R Documentation

Returns the table of vertices of a texture classification system.

Description

Returns the table of vertices of a texture classification system. Returns the clay silt sand coordinates of each vertices. Use TT.classes.tbl() to know the vertices that bounds each texture class. See also TT.vertices.plot().

Usage


TT.vertices.tbl(class.sys = "FAO50.TT")

Arguments

class.sys

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]