soiltexture-package(soiltexture)	R Documentation

Functions for soil texture plot classification and transformation

Description

"The Soil Texture Wizard" is a set of R functions designed to produce texture triangles (also called texture plots, texture diagrams, texture ternary plots), classify and transform soil textures data. These functions virtually allows to plot any soil texture triangle / classification into any triangle geometry (isosceles, right-angled triangles, etc.). This set of function is expected to be useful to people using soil textures data from different soil texture classification or different particle size systems. Several texture triangles are predefined: USDA; FAO (which is also the triangle for the soil map of Europe); Aisne (France); GEPPA (France); German triangle (Bodenkundliche Kartieranleitung 1994); Soil Survey of England and Wales (UK); Australian triangle; Belgian triangle; Canadian triangle; ISSS triangle; Romanian traingle. The Soil Texture Wizard is initially develloped from the functions 'soil.texture()' from the package 'plotrix' by Jim Lemon et al.

Details

Package: 	 soiltexture

Version: 	 1.02

Date: 	 2010-09-13

Title: 	 Functions for soil texture plot, classification and transformation

Author: 	 Julien MOEYS <jules_m78-soiltexture@yahoo.fr>, contributions from Wei Shangguan.

Maintainer: 	 Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

Depends: 	 R (>= 2.4.1), sp, MASS

Suggests: 	 drc

License: 	 GPL (>=3)

URL: 	 http://soiltexture.r-forge.r-project.org/

Author(s)

Julien MOEYS <jules_m78-soiltexture@yahoo.fr>

[Package soiltexture version 1.02 Index]